moving into ind components
