+++ frontend/src/App.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react';
 import './App.css';
-import DefaultD3 from './components/DefaultD3';
 import WbHeaderBsNav from './components/WbHeaderBsNav';
+import Retrieve from './components/Retrieve';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
         <WbHeaderBsNav />
-        <DefaultD3/>
+        <Retrieve />
       </div>
     );
   }
